requirements.txt
wheel
requests
python-magic
pillow
jinja2
pyenchant
pycryptodome
pytesseract
dulwich
beautifulsoup4
base58
pysocks
scipy
pydub
matplotlib
PyPDF2
pyopenssl
primefac
gmpy2
cmd2>=2.0.0
watchdog
pygments
regex
colorama
pyperclip
pwntools
